SubSegment Encode reference SUB_SEG_TYPES
Updated the _encode_segments function to reference the main list of SUB_SEG_TYPES instead of maintaining its own list.

Need to find a way to all the user to select which version of the spec they want to use for this.   0x30, 0x32, 0x44 and 0x46 were added to this in 2023r1 but implementations looking for 2022r1 freak out.